Artefatti/MongoDB/requirements.txt
mysql-connector-python>=8.0,<9.0
pymongo>=4.7.0
